Issue: #15 

- Utilize picture, source tags to render images in webp, jpeg and original as a fallback
- Modified image request to reduce size of the image
- Added lazy loading and fetch priorities on images to help speed up load times
- Refactored image data to clean up the file
- Refactored data fetch calls by limiting a call to only when API data provided is updated from what's currently in state.
- Lighthouse results:
   - Performance from 21 to 51 -- 30% increase
   - First contentful paint .3 to .2 -- .1s decrease
   - Speed index - 7.8s to 2s -- 5.8s decrease
   - Largest contentful paint - 3.4s to 2.2s -- 1.2s decrease
   - Cumulatative layout shift - .478 to .927 slight increase
  
<img width="1731" alt="Screenshot 2024-02-05 at 8 27 26 PM" src="https://github.com/plang-psm/popflix-v2/assets/101952500/a150cc45-6e08-4080-9b02-ae9d26219a9b">