@plang-psm 2024-02-06
Side note: The image placeholder for data without images is currently not appearing due to the picture/source tags. Issue will be created to address.